Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 int main(){
-    float number;
-    scanf("%f",&number);
-    printf("You entered: %f",number);
+    double number;
+    scanf("%lf",&number);
+    printf("You entered: %4lf",number);
     return 0;
 }
